+++ src/pages/Result.tsx
@@ -6,7 +6,7 @@ import Progress from '../components/Progress';
 import Rank from '../components/Rank';
 import useFetch, { cityGroup } from '../hooks/useFetchDustInfo';
 import { type SidoDust } from '@/type';
-import { FINE_DUST, ULTRA_FINE_DUST } from '@/utils/constance';
+import { FINE_DUST, ULTRA_FINE_DUST } from '@/utils/constants';
 
 const Result = () => {
   const [sidoDust, setSidoDust] = useState<SidoDust[] | []>([]);
